* Remove /Mac/BackgroundReader.mm, since BackgroundReader was removed too (dead code). * Change lots of printf calls to logOutput.Print, a major one that still needs   to be done is the 'print' call from LUA. (luaB_print) * Fixed an uninitialized memory read in InfoConsole.cpp. * Reformatted FontTexture.cpp to use tabs instead of spaces. * Removed some debug output from CobInstance. * Factored a Printv function out of all LogOutput functions (like vprintf) * Other minor fixes/cleanup.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@5114 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Lua/LuaSyncedCtrl.cpp
@@ -1163,22 +1163,22 @@ int LuaSyncedCtrl::SetUnitWeaponState(lua_State* L)
 				const string key = lua_tostring(L, -2);
 				const float value = (float)lua_tonumber(L, -1);
 				if (key == "reloadstate") {
-					unit->weapons[weapon]->reloadStatus=value;
+					unit->weapons[weapon]->reloadStatus = (int)value;
 				}
 				else if (key == "reloadtime") {
-					unit->weapons[weapon]->reloadTime=value*GAME_SPEED;
+					unit->weapons[weapon]->reloadTime = (int)(value*GAME_SPEED);
 				}
 				else if (key == "accuracy") {
-					unit->weapons[weapon]->accuracy=value;
+					unit->weapons[weapon]->accuracy = value;
 				}
 				else if (key == "sprayangle") {
-					unit->weapons[weapon]->sprayangle=value;
+					unit->weapons[weapon]->sprayangle = value;
 				}
 				else if (key == "range") {
-					unit->weapons[weapon]->range=value;
+					unit->weapons[weapon]->range = value;
 				}
 				else if (key == "projectilespeed") {
-					unit->weapons[weapon]->projectileSpeed=value;
+					unit->weapons[weapon]->projectileSpeed = value;
 				}
 			}
 		}
